requirements.txt
Django==1.4
TumblrAPI==0.2.4
django-social-auth==0.6.7
django-tumblr-auth==0.0.1
httplib2==0.7.4
oauth==1.0.1
oauth2==1.5.211
python-openid==2.2.5
wsgiref==0.1.2
